[refactor] Extract StreamTwoInputProcessor factory method
